Improved JSON functionality
The old method read in a JSON file as a list and required conversion to a model object to use, then back to a list object and finally to a dictionary when written back to the JSON file.

This new method works directly with the dictionaries of the relevant models and makes for easier reading/writing/manipulating.

As such, all old list-type methods have been deprecated.